Added bash conditions for afterinstall too

+++ scripts/afterInstall.bash
@@ -5,6 +5,10 @@
 
 sudo mkdir -vp /var/www/html/ecommerce/config/ssl/
 sudo cp -r /etc/ssl/paytron/* /var/www/html/ecommerce/config/ssl/
-sudo chown -R 1000 /var/www/html/ecommerce/.tmp
-sudo rm -rf /var/www/html/ecommerce/.tmp/public/
+
+if [ -d /var/www/html/ecommerce/.tmp/ ]; then
+    sudo chown -R 1000 /var/www/html/ecommerce/.tmp/*
+    sudo rm -rf /var/www/html/ecommerce/.tmp/public/
+fi
+
 cd /var/www/html/ecommerce && sudo npm i && sudo su && pm2 startOrRestart pm2.json --env production --watch
